Fix input value - change to per side, not total!
I had a feeling that this exercise was a little too
difficult and it turns out I'd already done what
I was supposed to.

Removed Math.sqrt() method and changed condition
for generation to multiply input number by itself.

You'll never have a number of squares that can't
be divided by itself if the input number is the side!

Thank goodness.